Moved enum and struct to another file
