+++ SEProject/Enums/DifficultyLevel.cs
@@ -1,0 +1,9 @@
+namespace SEProject.Enums
+{
+    public enum DifficultyLevel
+    {
+        Easy,
+        Medium,
+        Hard
+    }
+}
